Fix too frequent yields.
I forgot that signed minus unsigned is unsigned and therefore pd_hal_wfe_ms() mistakenly yielded every time, slowing things down a lot. (To be honest I would have expected a compiler warning about that, but it turns out -Wtype-limits is not included in -Wall.)

+++ src/mphal.c
@@ -184,7 +184,7 @@ void pd_hal_wfe_ms(int timeout_ms) {
 	// to be a sleep function in the Playdate API) or return immediately for
 	// timeout 0 as from MICROPY_VM_HOOK_POLL.
 	int end = global_pd->system->getCurrentTimeMilliseconds() + timeout_ms;
-	if (end - updateEndDue >= 0) {
+	if (end - (int)updateEndDue >= 0) {
 		pdco_yield(PDCO_MAIN_ID);
 	}
 	else if (timeout_ms > 0) {
